complete refactor codes
